+++ src/preprocessing.py
@@ -2,7 +2,7 @@ import numpy as np
 
 def normalizar_imagenes(X):
     """ Normaliza las imágenes dividiendo todos los valores por 255 """
-    return X.astype(np.float32) / 255.0
+    return X.astype(np.float32) / np.float32(255.0)
 
 def one_hot_encode(y, num_classes=None):
     """
@@ -14,7 +14,7 @@ def one_hot_encode(y, num_classes=None):
     if num_classes is None:
         num_classes = np.max(y) + 1
 
-    one_hot = np.zeros((y.shape[0], num_classes))
+    one_hot = np.zeros((y.shape[0], num_classes), dtype=np.float32)
     one_hot[np.arange(y.shape[0]), y] = 1
 
     return one_hot
